Explicitly convert int64 to double.
GitOrigin-RevId: afdf35ee316ebb7e317921d4e4734582203783e3

+++ td/telegram/CallActor.cpp
@@ -371,7 +371,7 @@ Status CallActor::do_update_call(telegram_api::phoneCallWaiting &call) {
       call_state_.is_received = true;
       call_state_need_flush_ = true;
       int64 call_ring_timeout_ms = G()->shared_config().get_option_integer("call_ring_timeout_ms", 90000);
-      set_timeout_in(call_ring_timeout_ms * 0.001);
+      set_timeout_in(static_cast<double>(call_ring_timeout_ms) * 0.001);
     }
   }
 
@@ -454,7 +454,7 @@ void CallActor::on_begin_exchanging_key() {
   call_state_.type = CallState::Type::ExchangingKey;
   call_state_need_flush_ = true;
   int64 call_receive_timeout_ms = G()->shared_config().get_option_integer("call_receive_timeout_ms", 20000);
-  double timeout = call_receive_timeout_ms * 0.001;
+  double timeout = static_cast<double>(call_receive_timeout_ms) * 0.001;
   LOG(INFO) << "Set call timeout to " << timeout;
   set_timeout_in(timeout);
 }
@@ -632,7 +632,7 @@ void CallActor::try_send_request_query() {
   auto query = G()->net_query_creator().create(tl_query);
   state_ = State::WaitRequestResult;
   int64 call_receive_timeout_ms = G()->shared_config().get_option_integer("call_receive_timeout_ms", 20000);
-  double timeout = call_receive_timeout_ms * 0.001;
+  double timeout = static_cast<double>(call_receive_timeout_ms) * 0.001;
   LOG(INFO) << "Set call timeout to " << timeout;
   set_timeout_in(timeout);
   query->total_timeout_limit_ = max(timeout, 10.0);
