+++ td/telegram/Td.cpp
@@ -3129,8 +3129,9 @@ void Td::on_online_updated(bool force, bool send_update) {
     update_status_query_ = create_handler<UpdateStatusQuery>()->send(!is_online_);
   }
   if (is_online_) {
-    alarm_timeout_.set_timeout_in(ONLINE_ALARM_ID,
-                                  G()->shared_config().get_option_integer("online_update_period_ms", 210000) * 1e-3);
+    alarm_timeout_.set_timeout_in(
+        ONLINE_ALARM_ID,
+        static_cast<double>(G()->shared_config().get_option_integer("online_update_period_ms", 210000)) * 1e-3);
   } else {
     alarm_timeout_.cancel_timeout(ONLINE_ALARM_ID);
   }
